requirements.txt
numpy==1.23.4
pandas==1.5.0
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.4
six==1.16.0
SQLAlchemy==1.4.42
XlsxWriter==3.0.3
